#47 refactor: 가독성을 위한 Authentication 패키지 구조 분할
