+++ backend/src/main/java/moheng/auth/application/AuthService.java
@@ -1,6 +1,11 @@
 package moheng.auth.application;
 
-import moheng.auth.domain.*;
+import moheng.auth.domain.oauth.OAuthClient;
+import moheng.auth.domain.oauth.OAuthMember;
+import moheng.auth.domain.oauth.OAuthProvider;
+import moheng.auth.domain.oauth.OAuthUriProvider;
+import moheng.auth.domain.token.MemberToken;
+import moheng.auth.domain.token.TokenManager;
 import moheng.auth.dto.LogoutRequest;
 import moheng.auth.dto.RenewalAccessTokenRequest;
 import moheng.auth.dto.RenewalAccessTokenResponse;
