+++ backend/src/main/java/moheng/auth/domain/oauth/KakaoUriProvider.java
@@ -1,4 +1,4 @@
-package moheng.auth.domain;
+package moheng.auth.domain.oauth;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
