+++ backend/src/main/java/moheng/auth/domain/oauth/OAuthMember.java
@@ -1,4 +1,4 @@
-package moheng.auth.domain;
+package moheng.auth.domain.oauth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
