+++ backend/src/main/java/moheng/auth/domain/oauth/OAuthProvider.java
@@ -1,4 +1,4 @@
-package moheng.auth.domain;
+package moheng.auth.domain.oauth;
 
 public interface OAuthProvider {
     OAuthClient getOauthClient(final String providerName);
