+++ backend/src/main/java/moheng/auth/domain/token/InMemoryRefreshTokenRepository.java
@@ -1,9 +1,8 @@
-package moheng.auth.domain;
+package moheng.auth.domain.token;
 
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
 @Component
