+++ backend/src/main/java/moheng/auth/domain/token/JwtTokenManager.java
@@ -1,14 +1,7 @@
-package moheng.auth.domain;
+package moheng.auth.domain.token;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import moheng.auth.exception.InvalidTokenException;
 import moheng.auth.exception.NoExistMemberTokenException;
 import org.springframework.stereotype.Component;
-
-import java.util.Date;
 
 @Component
 public class JwtTokenManager implements TokenManager {
