+++ backend/src/main/java/moheng/auth/domain/token/TokenManager.java
@@ -1,4 +1,4 @@
-package moheng.auth.domain;
+package moheng.auth.domain.token;
 
 public interface TokenManager {
     MemberToken createMemberToken(final long memberId);
