+++ backend/src/main/java/moheng/auth/presentation/admin/AdminAuthenticationArgumentResolver.java
@@ -1,7 +1,7 @@
 package moheng.auth.presentation.admin;
 
 import jakarta.servlet.http.HttpServletRequest;
-import moheng.auth.domain.JwtTokenProvider;
+import moheng.auth.domain.token.JwtTokenProvider;
 import moheng.auth.dto.Accessor;
 import moheng.auth.presentation.authentication.AuthenticationBearerExtractor;
 import org.springframework.core.MethodParameter;
