+++ backend/src/test/java/moheng/config/TestConfig.java
@@ -1,8 +1,6 @@
 package moheng.config;
 
-import moheng.auth.domain.OAuthClient;
-import moheng.auth.domain.OAuthClientProvider;
-import moheng.auth.domain.OAuthProvider;
+import moheng.auth.domain.oauth.OAuthProvider;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
